AI Wandering & AI Function Script
Call general AI scripts from AI Functions, update function for agent currently built for
