+++ Assets/Scripts/Objects/AI Scripts.meta
@@ -1,0 +1,8 @@
+fileFormatVersion: 2
+guid: fbdbea42d59f0454b874d3763cf59a1d
+folderAsset: yes
+DefaultImporter:
+  externalObjects: {}
+  userData: 
+  assetBundleName: 
+  assetBundleVariant:
