+++ Assets/Scripts/Objects/AI Scripts/AIAgentScript.cs.meta
@@ -1,0 +1,11 @@
+fileFormatVersion: 2
+guid: 8114eeabf0d46094e89754278030a5a6
+MonoImporter:
+  externalObjects: {}
+  serializedVersion: 2
+  defaultReferences: []
+  executionOrder: 0
+  icon: {instanceID: 0}
+  userData: 
+  assetBundleName: 
+  assetBundleVariant:
