+++ Assets/Scripts/Objects/AI Scripts/AIFunctions.cs.meta
@@ -1,0 +1,11 @@
+fileFormatVersion: 2
+guid: 244812ac65562d6429857b4ba8ffc861
+MonoImporter:
+  externalObjects: {}
+  serializedVersion: 2
+  defaultReferences: []
+  executionOrder: 0
+  icon: {instanceID: 0}
+  userData: 
+  assetBundleName: 
+  assetBundleVariant:
